+++ src/main/java/org/kumoricon/registration/settings/SettingsRepository.java
@@ -16,7 +16,9 @@ public class SettingsRepository {
     public static final String ENABLE_PRINTING = "enablePrinting";
     public static final String REPORT_PRINTER_NAME = "reportPrinterName";
     public static final String REQUIRE_STAFF_PHOTO = "requireStaffPhoto";
-    public static final String REQUIRE_STAFF_SIGNATURE= "requireStaffSignature";
+    public static final String REQUIRE_STAFF_SIGNATURE = "requireStaffSignature";
+    public static final String DEFAULT_PASSWORD = "defaultPassword";
+    public static final String FORCE_PASSWORD_CHANGE = "forcePasswordChange";
 
     public SettingsRepository(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
@@ -37,6 +39,7 @@ public class SettingsRepository {
                     return data;
                 });
 
+        if (results == null) return;
         if (results.containsKey(ENABLE_PRINTING)) {
             builder.setEnablePrinting(Boolean.parseBoolean(results.get(ENABLE_PRINTING)));
         }
@@ -45,6 +48,18 @@ public class SettingsRepository {
         }
         if (results.containsKey(TRAINING_MODE)) {
             builder.setTrainingMode(Boolean.parseBoolean(results.get(TRAINING_MODE)));
+        }
+        if (results.containsKey(DEFAULT_PASSWORD)) {
+            builder.setDefaultPassword(results.getOrDefault(DEFAULT_PASSWORD, "password"));
+        }
+        if (results.containsKey(FORCE_PASSWORD_CHANGE)) {
+            builder.setForcePasswordChange(Boolean.parseBoolean(results.get(FORCE_PASSWORD_CHANGE)));
+        }
+        if (results.containsKey(REQUIRE_STAFF_PHOTO)) {
+            builder.setRequireStaffPhoto(Boolean.parseBoolean(results.get(REQUIRE_STAFF_PHOTO)));
+        }
+        if (results.containsKey(REQUIRE_STAFF_SIGNATURE)) {
+            builder.setRequireStaffSignature(Boolean.parseBoolean(results.get(REQUIRE_STAFF_SIGNATURE)));
         }
     }
 }